create separate test suite
