+++ test/e2e/autoscale_test.go
@@ -443,16 +443,16 @@ func TestGracefulScaledown(t *testing.T) {
 		}))
 	defer test.TearDown(ctx.clients, ctx.names)
 
-	autoscalerConfigMap, err := rawCM(ctx.clients, autoscalerconfig.ConfigName)
+	autoscalerConfigMap, err := RawCM(ctx.clients, autoscalerconfig.ConfigName)
 	if err != nil {
 		t.Errorf("Error retrieving autoscaler configmap: %v", err)
 	}
 
 	patchedAutoscalerConfigMap := autoscalerConfigMap.DeepCopy()
 	patchedAutoscalerConfigMap.Data["enable-graceful-scaledown"] = "true"
-	patchCM(ctx.clients, patchedAutoscalerConfigMap)
-	defer patchCM(ctx.clients, autoscalerConfigMap)
-	test.CleanupOnInterrupt(func() { patchCM(ctx.clients, autoscalerConfigMap) })
+	PatchCM(ctx.clients, patchedAutoscalerConfigMap)
+	defer PatchCM(ctx.clients, autoscalerConfigMap)
+	test.CleanupOnInterrupt(func() { PatchCM(ctx.clients, autoscalerConfigMap) })
 
 	if err = assertGracefulScaledown(t, ctx, 4 /* desired pods */); err != nil {
 		t.Errorf("Failed: %v", err)
